backend_app/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.8
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
rpds-py==0.18.1
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
uWSGI==2.0.26